Configuration Provider Cancellation Refactor (#78)
* Add Configuration Overloads without Cancellation

Updates the cofiguration builder extensions to allow adding the consul
configuration provider without specifying a cancellation token. If no
token is provided then the default cancellation token is used instead.

* Cancel Watching if Config Provider Disposed

Add an `IDisposable` implemnetation to the `ConsulConfigurationProvider`.
When disposed the provider cancels  the watch of the KV node. This
brings the cancellation behavior more in line with that of other config
providers.

There are still a few open questions with this however:

 * The cancellation token source _and_ chnage token registration are a
    bit clunky. Ideally we could just have one or the other but the
    change token doesn't cancel the watch properly if that's all we
    have.
 * Ideally if we _do_ need two cancellation tokens, one from the source
    and one to pass along the `Dispose` call we would combine them with
    `CancellationTokenSource.CreateLinkedTokenSource` rather than the
    registration we have right now. This may work better if the watching
    of the KV node is moved more into the configuration provider.

* Remove `CancellationToken` from `ConsulConfigurationSource`

With the `IDisposable` support for the `ConsulConfiugrationProvider`
this token is no longer needed.

semver: breaking.

* Fixup Unit Test Mocking

Switch to accepting any cancellation token to fixup the failing test
cases

* Switch to Task for Polling Changes

Move away from using change tokens to provide better `async` support
when checking for changes. This brings the polling for changes more in
line with other config providers with background polling task.

* Fixup Documentation Violations

* Address Race in Reload Test

Fixup the race between the test and the background task performing a
reload.

* Make `PollWaitTime` Configurable

Addresses the confugrability of the wait when polling for
changes. Asrequested in PR comments.

* Extract Polling into Separate Method

This runs the first part of the poll within the context of the
constructor to bring the behaviour back in line with the origional
changetoken based polling.

* Code Review Fixups

Address issues from PR review.

* Refactor to remove ConsulConfigurationClient class.

* Re-add net461 and add test for default KeyToRemove.

* Fix comments.