add IConfigurationKeyValueRewriter interface
allows IConfigurationParser implementations to opt-in to the ability
to rewrite key/value configuration pairs before they get added to
the configuration.

example may be stripping underscores, rewriting double underscores
to / or : to match the behaviour of Microsoft.Extensions.Configuration
when parsing environment variables, or intercepting tokens from the
value to inject secret values which are not stored in Consul.

NOTE: this commit also unseals SimpleConfigurationParser and
JsonConfigurationParser so that they can be used as base classes for
clients who want this default behaviour but want to add additional
opt-in support for rewriting configuration key/values.