rewrite ConvertToConfig using composition
rearrange the processing to compose helper functions using
Select(...); the helper functions help decompose the individual
processing steps of the ConvertToConfig(...) pipeline.

this also separates the normalizing of config keys from the removal
of a key prefix to ensures that IConfigurationKeyValueRewriters
get to operate on fully qualified and normalized keys.